+++ src/test/java/globallogic/evaluate/msglsignup/controller/UserControllerMvcTest.java
@@ -15,7 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-class SignInControllerMvcTest {
+class UserControllerMvcTest {
 
     @Autowired
     private MockMvc mockMvc;
