+++ src/test/java/globallogic/evaluate/msglsignup/controller/UserControllerTest.java
@@ -1,10 +1,13 @@
 package globallogic.evaluate.msglsignup.controller;
 
+import globallogic.evaluate.msglsignup.dto.CreateUserDto;
 import globallogic.evaluate.msglsignup.dto.GetAccessTokenDto;
 import globallogic.evaluate.msglsignup.dto.SignInDto;
+import globallogic.evaluate.msglsignup.exception.MailAlreadyRegisteredException;
 import globallogic.evaluate.msglsignup.exception.UserNotFoundException;
 import globallogic.evaluate.msglsignup.security.jwt.JwtProvider;
 import globallogic.evaluate.msglsignup.service.SignInService;
+import globallogic.evaluate.msglsignup.service.SignUpService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -16,6 +19,7 @@ import org.springframework.security.authentication.AuthenticationManager;
 
 import java.util.UUID;
 
+import static globallogic.evaluate.msglsignup.DataMock.createUserMock;
 import static globallogic.evaluate.msglsignup.DataMock.getUserMock;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -25,10 +29,10 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class SignInControllerTest {
+class UserControllerTest {
 
     @InjectMocks
-    private SignInController signInController;
+    private UserController userController;
 
     @Mock
     private JwtProvider jwtProvider;
@@ -39,13 +43,16 @@ class SignInControllerTest {
     @Mock
     private SignInService signInService;
 
+    @Mock
+    private SignUpService signUpService;
+
     @Test
     @DisplayName("Successful SignIn")
     void signInTest() {
         var accessToken = "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJkdW1teUBnbWFpbC5jb20iLCJpc3MiOiJnbG9iYWwgbG9naWMiLCJpYXQiOjE3MTc5NzU5MDksImV4cCI6MTcxNzk4MzEwOX0.y_TcHzcqRZPGs0Yz8mZzt8BQrYIYKUMs4O7I6bsMCOA";
         when(signInService.signIn(any(SignInDto.class)))
                 .thenReturn(new GetAccessTokenDto(accessToken));
-        var response = signInController.signIn(new SignInDto("", ""));
+        var response = userController.signIn(new SignInDto("", ""));
         assertAll(
                 () -> assertThat(response.getStatusCode(), equalTo(HttpStatus.OK)),
                 () -> assertThat(response.getBody(), equalTo(new GetAccessTokenDto(accessToken)))
@@ -58,7 +65,7 @@ class SignInControllerTest {
         var signInDto = new SignInDto("dummy@gmail.com", "Password12");
         when(signInService.signIn(signInDto)).thenThrow(UserNotFoundException.class);
         assertThrows(UserNotFoundException.class,
-                () -> signInController.signIn(signInDto));
+                () -> userController.signIn(signInDto));
     }
 
     @Test
@@ -67,10 +74,31 @@ class SignInControllerTest {
         var getUserDto = getUserMock();
         when(signInService.getUserDetailById(any(UUID.class)))
                 .thenReturn(getUserDto);
-        var response = signInController.getUserDetail(UUID.randomUUID());
+        var response = userController.getUserDetail(UUID.randomUUID());
         assertAll(
                 () -> assertThat(response.getStatusCode(), equalTo(HttpStatus.OK)),
                 () -> assertThat(response.getBody(), equalTo(getUserDto))
         );
     }
+
+    @Test
+    @DisplayName("Save new user")
+    void saveNewUserOkTest() {
+        when(signUpService.saveNewUser(any(CreateUserDto.class)))
+                .thenReturn(getUserMock());
+        var response = userController.createNewUser(createUserMock());
+        assertAll(
+                () -> assertThat(response.getStatusCode(), equalTo(HttpStatus.CREATED)),
+                () -> assertThat(response.getBody(), equalTo(getUserMock()))
+        );
+    }
+
+    @Test
+    @DisplayName("Save user with already registered mail")
+    void alreadyRegisteredEmailTest() {
+        when(signUpService.saveNewUser(any(CreateUserDto.class)))
+                .thenThrow(MailAlreadyRegisteredException.class);
+        assertThrows(MailAlreadyRegisteredException.class, () -> userController.createNewUser(CreateUserDto.builder().build()));
+    }
+
 }
